requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
opencv-python==4.12.0.88
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
regex==2025.7.34
requests==2.32.4
sympy==1.13.1
tiktoken==0.11.0
torch==2.5.1+cu121
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
